fixxed compile issue

+++ ft_chrdindex.c
@@ -6,13 +6,13 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/02/03 22:38:37 by pholster       #+#    #+#                */
-/*   Updated: 2019/05/02 19:13:48 by pholster      ########   odam.nl         */
+/*   Updated: 2019/05/02 19:22:22 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/libft.h"
 
-size_t		ft_chrdindex(const char *str, const char *dlm)
+int		ft_chrdindex(const char *str, const char *dlm)
 {
 	size_t		i;
 	long long	bytes[2];
@@ -24,7 +24,7 @@ size_t		ft_chrdindex(const char *str, const char *dlm)
 	while (str[i] != '\0')
 	{
 		if (ft_chrinbytes(bytes, str[i]))
-			return (i);
+			return ((int)i);
 		i++;
 	}
 	return (-1);
